+++ src/pages/detailed/DetailedProductPage.tsx
@@ -5,7 +5,6 @@ import {
   Grid,
   SimpleGrid,
   Paper,
-  Button,
   Title,
   Group,
   Text,
@@ -14,20 +13,17 @@ import {
   rem,
   getStylesRef,
   createStyles,
-  Select,
 } from '@mantine/core';
 import { Carousel } from '@mantine/carousel';
 import ModalCarousel from '../../components/modal-carousel/ModalCarousel';
-import { ProductProjection, ProductVariant } from '@commercetools/platform-sdk';
+import { ProductProjection } from '@commercetools/platform-sdk';
 import { storeService } from '../../services/StoreService/StoreService';
 import { ErrorCodes, getErrorMessage } from '../../utils/helpers/error-handler';
-import { notificationError, notificationSuccess } from '../../components/ui/notification';
-import useAuth from '../../utils/hooks/useAuth';
-import { PiBagSimple } from 'react-icons/pi';
+import { notificationError } from '../../components/ui/notification';
 import parse from 'html-react-parser';
-import { checkProductInCart } from '../../utils/helpers/cart-helpers';
 import { formatPrice } from '../../utils/helpers/format-price';
 import './detailed-product-page.scss';
+import CartSelector from '../../components/cart-selector/CartSelector';
 
 const carouselStyles = createStyles((theme) => ({
   root: {
@@ -68,14 +64,6 @@ const carouselStyles = createStyles((theme) => ({
   },
 }));
 
-const getSizeData = (variant: ProductVariant) => ({
-  label: variant.attributes?.find((attribute) => attribute.name === 'size')?.value.label,
-  value: `${variant.id}`,
-});
-const getProductSizes = (product: ProductProjection) => {
-  return [getSizeData(product.masterVariant), ...product.variants.map((variant) => getSizeData(variant))];
-};
-
 type DetailedProductPageProps = {
   isLoading: boolean;
   setIsLoading: React.Dispatch<React.SetStateAction<boolean>>;
@@ -83,14 +71,11 @@ type DetailedProductPageProps = {
 
 const DetailedProductPage = ({ isLoading, setIsLoading }: DetailedProductPageProps): JSX.Element => {
   const [product, setProduct] = useState<ProductProjection>();
-  const [selectedVariant, setSelectedVariant] = useState<string | null>(null);
-  const [buttonDisabled, setButtonDisabled] = useState(false);
   const { productID } = useParams();
   const { classes } = carouselStyles();
   const navigate = useNavigate();
   const [opened, setOpened] = useState(false);
   const [initialSlide, setInitialSlide] = useState(0);
-  const { cart, setCart } = useAuth();
 
   const slides = (product: ProductProjection): JSX.Element[] | null => {
     if (product.masterVariant.images) {
@@ -121,11 +106,6 @@ const DetailedProductPage = ({ isLoading, setIsLoading }: DetailedProductPagePro
     };
     fetchData();
   }, [productID, navigate]);
-
-  useEffect(() => {
-    if (product && selectedVariant && cart)
-      setButtonDisabled(() => checkProductInCart(product?.id, +selectedVariant, cart));
-  }, [selectedVariant, product, cart]);
 
   return (
     <Container w="100%" h="100%" my="md" px="1rem" size="lg">
@@ -159,67 +139,8 @@ const DetailedProductPage = ({ isLoading, setIsLoading }: DetailedProductPagePro
                       {product.masterVariant.prices && product.masterVariant.prices[0].value.centAmount / 100 + ' €'}
                     </Text>
                   </Group>
-                  <Select
-                    my="md"
-                    mr="md"
-                    withinPortal
-                    data={getProductSizes(product)}
-                    onChange={setSelectedVariant}
-                    placeholder="Select size"
-                  />
-                  <Group spacing={5}>
-                    <Button
-                      loading={isLoading && !buttonDisabled}
-                      rightIcon={<PiBagSimple size="1.5rem" />}
-                      mr="sm"
-                      onClick={async () => {
-                        if (!selectedVariant) return;
-                        setIsLoading(true);
-                        try {
-                          const updatedCart = await storeService.addProductToCart(product.id, +selectedVariant);
-                          if (updatedCart) {
-                            notificationSuccess('Item added to cart');
-                            setCart(updatedCart);
-                            setButtonDisabled(true);
-                          }
-                        } catch (err) {
-                          if (err instanceof Error) notificationError(err.message);
-                        } finally {
-                          setIsLoading(false);
-                        }
-                      }}
-                      disabled={buttonDisabled}
-                    >
-                      Add To
-                    </Button>
-                    {buttonDisabled && (
-                      <Button
-                        loading={isLoading}
-                        rightIcon={<PiBagSimple size="1.5rem" />}
-                        onClick={async () => {
-                          if (cart && selectedVariant) {
-                            setIsLoading(true);
-                            try {
-                              let updatedCart = await storeService.removeProductFromCart(product.id, +selectedVariant);
-                              if (updatedCart?.lineItems.length === 0) {
-                                await storeService.deleteCart();
-                                updatedCart = null;
-                              }
-                              notificationSuccess('Item removed from cart');
-                              setCart(updatedCart);
-                              setButtonDisabled(false);
-                            } catch (err) {
-                              if (err instanceof Error) notificationError(err.message);
-                            } finally {
-                              setIsLoading(false);
-                            }
-                          }
-                        }}
-                      >
-                        Remove From
-                      </Button>
-                    )}
-                  </Group>
+
+                  <CartSelector product={product} isLoading={isLoading} setIsLoading={setIsLoading} />
 
                   <Paper fz={13} className="product-description" ff="Montserrat">
                     {product.description?.['en-US'] && parse(product.description['en-US'])}
